requirements.txt
python-socketio
python-socketio[client]
aiohttp
